requirements.txt
crewai
crewai[tools]
streamlit
faiss-cpu
python-dotenv